Improve rtdn api usage.
