+++ gradle/libs.versions.toml
@@ -16,7 +16,7 @@ androidx_preference_version = "1.2.0"
 androidx_room_version = "2.6.0-beta01"
 androidx_splashscreen = "1.0.1"
 build_tools_version = "30.0.3"
-catappult_billing_version = "0.9.0.4"
+catappult_billing_version = "0.9.0.5.dev-SNAPSHOT"
 coil_version = "2.3.0"
 corektx_version = "1.12.0"
 coroutines_version = "1.7.2"
