+++ payments/appcoins-sdk/src/main/java/com/appcoins/diceroll/sdk/payments/appcoins_sdk/SdkManager.kt
@@ -76,7 +76,7 @@ interface SdkManager {
                             queryPurchases()
                             queryActiveSubscriptions()
                             queryInappsSkus(ArrayList(listOf("attempts")))
-                            querySubsSkus(ArrayList(listOf("golden_dice")))
+                            querySubsSkus(ArrayList(listOf("daily_dice")))
                         }
 
                         else -> {
@@ -195,7 +195,7 @@ interface SdkManager {
                     if (sku.sku == "attempts") {
                         _attemptsPrice.value = sku.price
                     }
-                    if (sku.sku == "golden_dice") {
+                    if (sku.sku == "daily_dice") {
                         _goldDicePrice.value = sku.price
                     }
                     // You can add these details to a list in order to update
@@ -281,7 +281,7 @@ interface SdkManager {
 
     private fun processPurchase(purchase: Purchase) {
         when (purchase.sku) {
-            "golden_dice" -> processGoldenDiceSubscription(true)
+            "daily_dice" -> processGoldenDiceSubscription(true)
         }
     }
 
@@ -304,7 +304,7 @@ interface SdkManager {
                 _purchases.add(purchase)
                 validateAndConsumePurchase(purchase, true)
             }
-            if (purchases.find { it.sku == "golden_dice" } == null) {
+            if (purchases.find { it.sku == "daily_dice" } == null) {
                 processGoldenDiceSubscription(false)
             }
         }
