+++ payments/appcoins-sdk/src/main/java/com/appcoins/diceroll/sdk/payments/appcoins_sdk/SdkManagerImpl.kt
@@ -2,12 +2,12 @@ package com.appcoins.diceroll.sdk.payments.appcoins_sdk
 
 import android.content.Context
 import android.util.Log
+import android.widget.Toast
 import com.appcoins.diceroll.sdk.core.network.clients.RTDNWebSocketClient
 import com.appcoins.diceroll.sdk.feature.roll_game.data.usecases.GetGoldenDiceStatusUseCase
 import com.appcoins.diceroll.sdk.feature.roll_game.data.usecases.UpdateGoldenDiceStatusUseCase
 import com.appcoins.diceroll.sdk.payments.appcoins_sdk.SdkManager.Companion.LOG_TAG
 import com.appcoins.diceroll.sdk.payments.appcoins_sdk.data.respository.PurchaseValidatorRepository
-import com.appcoins.diceroll.sdk.payments.appcoins_sdk.usecases.ProcessMessageReceivedFromRTDNUseCase
 import com.appcoins.sdk.billing.AppcoinsBillingClient
 import com.appcoins.sdk.billing.Purchase
 import com.appcoins.sdk.billing.helpers.CatapultBillingAppCoinsFactory
@@ -17,6 +17,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.launch
+import org.json.JSONObject
 import javax.inject.Inject
 
 /**
@@ -37,8 +38,7 @@ class SdkManagerImpl @Inject constructor(
     purchaseValidatorRepository: PurchaseValidatorRepository,
     val getGoldenDiceStatusUseCase: GetGoldenDiceStatusUseCase,
     val updateGoldenDiceStatusUseCase: UpdateGoldenDiceStatusUseCase,
-    private val webSocketClient: RTDNWebSocketClient,
-    val processMessageReceivedFromRTDNUseCase: ProcessMessageReceivedFromRTDNUseCase
+    private val webSocketClient: RTDNWebSocketClient
 ) : SdkManager {
 
     override lateinit var cab: AppcoinsBillingClient
@@ -94,6 +94,44 @@ class SdkManagerImpl @Inject constructor(
     private val rtdnListener: (String) -> Unit
         get() = { message ->
             Log.i(LOG_TAG, "Received RTDN message: $message")
-            processMessageReceivedFromRTDNUseCase(message)
+            processMessageFromRTDN(message)?.let { messageToShow ->
+                CoroutineScope(Dispatchers.Main).launch {
+                    Toast.makeText(context, messageToShow, Toast.LENGTH_LONG).show()
+                }
+            }
+        }
+
+    private fun processMessageFromRTDN(message: String): String? {
+        try {
+            val jsonObject = JSONObject(message)
+            return when (val sku = jsonObject.optString("sku")) {
+                "attempts" -> processAttemptsPurchaseUpdate(jsonObject)
+                "daily_dice" -> processGoldenDiceSubscriptionUpdate(jsonObject)
+                else -> "Unknown SKU received. $sku"
+            }
+        } catch (ex: Exception) {
+            Log.e(LOG_TAG, "Failed to parse message from RTDN.", ex)
+        }
+
+        return "Couldn't parse the message from RTDN."
+    }
+
+    private fun processAttemptsPurchaseUpdate(jsonObject: JSONObject): String? =
+        if (jsonObject.optString("status") == "VOIDED") {
+            "Your purchase for more Attempts was voided."
+        } else {
+            Log.i(LOG_TAG, "Status is not important for the Attempts purchase.")
+            null
+        }
+
+    private fun processGoldenDiceSubscriptionUpdate(jsonObject: JSONObject): String? =
+        if (jsonObject.optString("status") == "VOIDED") {
+            CoroutineScope(Dispatchers.IO).launch {
+                processGoldenDiceSubscription(false)
+            }
+            "Your subscription for the Golden Dice was voided."
+        } else {
+            Log.i(LOG_TAG, "Status is not important for the GoldenDice purchase.")
+            null
         }
 }
